+++ src/src.pro
@@ -6,5 +6,9 @@ QT = core gui
 greaterThan(QT_MAJOR_VERSION, 4): QT += widgets
 
 SOURCES += \
-    backend/expression.cpp \
+    backend/plotter.cpp \
     gui/gui.cpp
+
+LIBS+= -lGL -lGLU -lglut
+
+HEADERS +=
